+++ lib/test/utils/transform_trackingnet.py
@@ -20,8 +20,8 @@ def transform_trackingnet(tracker_name, cfg_name):
         if "time" not in item:
             src_path = os.path.join(src_dir, item)
             dest_path = os.path.join(dest_dir, item)
-            bbox_arr = np.loadtxt(src_path, dtype=np.int, delimiter='\t')
-            np.savetxt(dest_path, bbox_arr, fmt='%d', delimiter=',')
+            bbox_arr = np.loadtxt(src_path, dtype=np.float_, delimiter='\t')
+            np.savetxt(dest_path, bbox_arr, fmt='%.9f', delimiter=',')
     # make zip archive
     shutil.make_archive(src_dir, "zip", src_dir)
     shutil.make_archive(dest_dir, "zip", dest_dir)
